requirements.txt
fastapi
python-dotenv
httpx
streamlit